django_api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
numpy>=1.26.0

